Added toggleable Disc display to output
Toggle to display the discs of enemies in the output (the table holding them will also be adjusted), as well as the maximum number of actions that enemy can take per turn, if it differs from the number of their displayed discs. Enemies with only one type of disc and fewer than three actions will only display discs up to the number of actions they can take.
Also, added a button to clear the input field.