Added Field Effects functionality
The script will now automatically parse field effects, and format them into  FieldEffectsBox and FieldEffects templates. With the current implementation, an effect affecting multiple attributes (but not All) will only correctly compile if each attribute instance of the effect is listed consecutively in the JSON (as it is for Elsa Maria). If they are not, a separate instance of the effect will be listed for each attribute. This should be fairly easy to fix by hand, but would be substantially more difficult to code an automatic solution for than the current implementation.

Also, added Elsa Maria to characters.json, and removed some superfluous code of mine.